ANIM-RT truncation fixes - AnimationBlueprintLibrary #jira UE-166629 #rb jaime.cifuentes #preflight 63512aac8449f85a47b5fa16
[CL 22652656 by Jurre deBaare in ue5-main branch]

+++ Engine/Source/Editor/AnimationBlueprintLibrary/Private/AnimationBlueprintLibrary.cpp
@@ -2383,7 +2383,7 @@ bool UAnimationBlueprintLibrary::EvaluateRootBoneTimecodeAttributesAtTime(const 
 		}
 		else
 		{
-			FrameRate = FFrameRate(FMath::TruncToInt(TimecodeRateAsDecimal), 1);
+			FrameRate = FFrameRate(static_cast<int32>(FMath::TruncToInt(TimecodeRateAsDecimal)), 1);
 		}
 	}
 	else if (const UAnimSequence* AnimSequence = AnimDataModel->GetAnimationSequence())
